Revert "Commit 7c130ee"
This PR reverts commit 7c130ee96e1ac003ce2b22ba51cedab1ca119598, which was accidentally submitted.

Original MR description: 
Only check for the presence of apiSubDomain, i.e. "model", as opposed to apiSubDomain+".", i.e. "model.", when determining whether the URL needs to be modified.
This allows URLs such as `model-v1alpha.{domain}` to be unmodified by this logic, which is the desired behavior.